+++ SVRichPresence/CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added a check for Android, which will cancel mod initialization.
+
+### Changed
+- Small change to the debug build warning.
+
+### Fixed
+- Finally fixed Linux and Mac support and re-added the auto setup.
 
 ## [2.3.0-beta1] - 2019-05-12
 ### Added